migrate product view
